+++ AAEmu.Game/Models/Game/Units/Mate.cs
@@ -59,7 +59,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -82,7 +82,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -105,7 +105,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -128,7 +128,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -151,7 +151,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -171,7 +171,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -230,7 +230,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -261,7 +261,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -293,7 +293,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -324,7 +324,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -355,7 +355,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
 
             return res;
@@ -407,7 +407,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += res * bonus.Value / 100f;
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
             return (int)res;
         }
@@ -436,7 +436,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += res * bonus.Value / 100f;
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
             return (int)res;
         }
@@ -464,7 +464,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
             return res;
         }
@@ -492,7 +492,7 @@ public sealed class Mate : Unit
                 if (bonus.Template.ModifierType == UnitModifierType.Percent)
                     res += (int)(res * bonus.Value / 100f);
                 else
-                    res += bonus.Value;
+                    res += (int)bonus.Value;
             }
             return res;
         }
